look AT THE TIMESTAMP OF THIS GITHUB ACTIVITY
ahh, this is keeping me up at night 

Nightmare of History I a trying to Awake - James Joyce

Minute 22 of Episode One. Please, bypass the mindset of censoring as "boring" and non-attractive. Joyce speaks!

Published year 1988

MOYERS: You have written that “the sign of the cross has to be looked upon as a sign of an eternal affirmation of all that ever was or shall ever be. It symbolizes not only the one historic moment on Calvary but the mystery through all time and space of God’s presence and participation in the agony of all living things.”
CAMPBELL: The big moment in the medieval myth is the awakening of the heart to compassion, the transformation of passion into compassion. That is the whole problem of the Grail stories, compassion for the wounded king. And out of that you also get the notion that Abelard offered as an explanation of the crucifixion: that the Son of God came down into this world to be crucified to awaken our hearts to compassion, and thus to turn our minds from the gross concerns of raw life in the world to the specifically human values of self-giving in shared suffering. In that sense the wounded king, the maimed king of the Grail legend, is a counterpart of the Christ. He is there to evoke compassion and thus bring a dead wasteland to life. There is a mystical notion there of the spiritual function of suffering in this world. The one who suffers is, as it were, the Christ, come before us to evoke the one thing that turns the human beast of prey into a valid human being. That one thing is compassion. This is the theme that James Joyce takes over and develops in Ulysses